+++ docs/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0/).
 
+## [0.2.4] - 2025-12-17
+
+### Changed
+- Rebranded "Events Calendar" to "Live Music Calendar" across homepage and archive titles
+- Added archive title display on taxonomy archive pages using DataMachineEvents Archive_Title class
+- Added datamachine_events_archive_title filter to customize archive titles with "Live Music Calendar" branding
+
 ## [0.2.3] - 2025-12-16
 
 ### Added